Add files via upload
Added a jupyter notebook file which can be used to import force and distance data from .h5 kymograph files to a pandas dataframe and crop the dataframe to a selected time window via an interactive plot. An additional dataframe column containing smoothed values for a selected data channel is added, and the parameters for this smoothing Butterworth filter are also adjustable via the interactive plot. These filtered values are often better suited to downstream analysis – especially when derivatives need to be calculated.